// apps/mobile/src/services/tradeCore/account/typings.d.ts
declare namespace Account {
  // 账号交易品种及配置-集合-参数
  type TradeSymbolListParams = {
    /**
     * 交易账户ID
     */
    accountId?: string | number
    /**
     * 交易品种
     */
    symbol?: string
    /**
     * 交易品种路径 例如 /test2/1/*
     */
    symbolPath?: string
    /** 品种分类参数 */
    classify?: string
  }
  // 账号交易品种及配置-集合-列表
  type TradeSymbolListItem = {
    /*
     * 该字段渲染列表用
     */
    visible?: boolean
    /**
     * 别名
     */
    alias?: string
    /**
     * 数据源code
     */
    dataSourceCode?: string
    /**
     * 数据源Symbol
     */
    dataSourceSymbol?: string
    /**
     * 主键
     */
    id: number
    /**
     * 图标
     */
    imgUrl?: string
    /**
     * 备注
     */
    remark?: string
    /**
     * 状态
     */
    status?: API.Status
    /**
     * 品种名称
     */
    symbol: string
    symbolConf?: Symbol.SymbolConf
    /**
     * 品种组配置ID
     */
    symbolConfId?: number
    /**
     * 品种的价格小数位
     */
    symbolDecimal?: number
    /**
     * 品种组ID
     */
    symbolGroupId?: number
    /**
     * 自定义添加的是否选中
     */
    checked?: boolean
    /** 品种高开低收 */
    symbolNewTicker?: SymbolNewTicker
    /** 第一口报价信息 */
    symbolNewPrice?: SymbolNewPrice
    /** 账户组ID */
    accountGroupId?: string
    /** 品种分类编号 10加密货币 20大宗商品 30外汇 40指数 50股票 */
    classify?: string
    /** 自定义字段 是否在假期内 */
    isInHoliday?: boolean
  }
  type SymbolNewPrice = {
    /** 卖交易量 */
    sellSize: string
    /** 卖价 */
    sell: number
    /** 买价 */
    buy: number
    /** 时间戳 */
    id: number
    /** 买交易量 */
    buySize: string
  }
  // 高开低收，初始值
  type SymbolNewTicker = {
    dataSourceCode: string
    symbol: string
    /** 开盘价 */
    open: string
    /** 收盘价 */
    close: string
    /** 最低 */
    low: string
    /** 最高 */
    high: string
  }
// ...
}